% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{summarise,data.frame-method}
\alias{summarise,data.frame-method}
\title{Method for \code{summarise} on data.frame}
\usage{
\S4method{summarise}{data.frame}(object, ...)
}
\arguments{
\item{object}{A data.frame object.}

\item{...}{Additional arguments.}
}
\description{
Method for \code{summarise} on data.frame
}
