% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{summarise,grouped_df_list-method}
\alias{summarise,grouped_df_list-method}
\title{Method for \code{summarise} on grouped_df_list}
\usage{
\S4method{summarise}{grouped_df_list}(object, ...)
}
\arguments{
\item{object}{A grouped_df_list object.}

\item{...}{Additional arguments.}
}
\value{
An object of the same class as \code{.data.} One grouping level will be dropped.
}
\description{
Method for \code{summarise} on grouped_df_list
}
