% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{summarize,grouped_df_list-method}
\alias{summarize,grouped_df_list-method}
\title{Method for \code{summarize} on grouped_df_list}
\usage{
\S4method{summarize}{grouped_df_list}(object, ...)
}
\arguments{
\item{object}{A grouped_df_list object.}

\item{...}{Additional arguments.}
}
\description{
Method for \code{summarize} on grouped_df_list
}
