% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{summarize}
\alias{summarize}
\title{Generic \code{summarize} function}
\usage{
summarize(object, ...)
}
\arguments{
\item{object}{Object to be summarized.}

\item{...}{Additional arguments.}
}
\value{
An object of the same class as \code{.data}. One grouping level will be dropped.
}
\description{
Generic \code{summarize} function
}
