% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{summarize_rollup}
\alias{summarize_rollup}
\title{summarize_rollup}
\usage{
summarize_rollup(df_list, ...)
}
\arguments{
\item{df_list}{'grouped_df_list' class}

\item{...}{functions for 'summarize'}
}
\value{
An object of the same class as \code{.data}. The unioned aggregated result of multiple grouping levels will be dropped.
}
\description{
'summarize_rollup' aggregates each 'grouped_df' in the 'grouped_df_list' class and return the unioned aggregated results.
}
