% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{web_service_data}
\alias{web_service_data}
\title{Web Service Data}
\format{
A data frame with 30,000 rows and 6 variables:
\describe{
\item{date_id}{date id}
\item{id}{user id}
\item{gender}{gender}
\item{age}{age band}
\item{page_view_cnt}{pageview count}
\item{product_view_cnt_cat}{product view count (category)}
}
}
\source{
Generated for example purposes
}
\usage{
web_service_data
}
\description{
A dataset containing information about various web services.
}
\keyword{datasets}
