% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{with_rollup}
\alias{with_rollup}
\title{With Rollup for R dataframes}
\usage{
with_rollup(grouped_df)
}
\arguments{
\item{grouped_df}{}
}
\value{
list of dataframes
}
\description{
With Rollup for R dataframes
}
\examples{
mtcars \%>\% group_by(vs, am) \%>\% grouping_sets("vs","am",c("vs","am")) \%>\% summarize(avg_mpg = mean(mpg), avg_disp = mean(disp))
mtcars \%>\% group_by(vs, am) \%>\% with_rollup() \%>\% summarize(avg_mpg = mean(mpg), avg_disp = mean(disp))
mtcars \%>\% group_by(vs, am) \%>\% with_cube() \%>\% summarize(avg_mpg = mean(mpg), avg_disp = mean(disp)) 
}
