% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{grouping_sets}
\alias{grouping_sets}
\title{Grouping Sets for R dataframe}
\usage{
grouping_sets(df, ...)
}
\arguments{
\item{df}{dataframe or grouped df}

\item{...}{grouping variables}
}
\value{
list of dataframes grouped by grouping variables
}
\description{
Compute total amounts at different group levels, producing multiple subtotals. This mirrors the GROUPING SETS operations in SQL.
}
\examples{
mtcars \%>\% group_by(vs, am) \%>\% grouping_sets("vs","am",c("vs","am")) \%>\% summarize(avg_mpg = mean(mpg), avg_disp = mean(disp))
mtcars \%>\% group_by(vs, am) \%>\% with_rollup() \%>\% summarize(avg_mpg = mean(mpg), avg_disp = mean(disp))
mtcars \%>\% group_by(vs, am) \%>\% with_cube() \%>\% summarize(avg_mpg = mean(mpg), avg_disp = mean(disp)) 
}
