% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{summarise}
\alias{summarise}
\title{Generic \code{summarise} function}
\usage{
summarise(object, ...)
}
\arguments{
\item{object}{Object to be summarized.}

\item{...}{Additional arguments.}
}
\description{
Generic \code{summarise} function
}
