% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{summarize,ANY-method}
\alias{summarize,ANY-method}
\title{Default method for \code{summarize}}
\usage{
\S4method{summarize}{ANY}(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Additional arguments.}
}
\value{
An object of the same class as \code{.data}. One grouping level will be dropped.
}
\description{
Default method for \code{summarize}
}
