% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{summarize_rollup}
\alias{summarize_rollup}
\title{Summarize returns of Grouping Sets}
\usage{
summarize_rollup(df_list, ...)
}
\arguments{
\item{grouped_df_list}{}

\item{functions}{for 'summarize'}
}
\value{
summarized dataframe
}
\description{
Summarize returns of Grouping Sets
}
\examples{
mtcars \%>\% group_by(vs, am) \%>\% grouping_sets("vs","am",c("vs","am")) \%>\% summarize(avg_mpg = mean(mpg), avg_disp = mean(disp))
}
