% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{summarize_rollup}
\alias{summarize_rollup}
\title{Summarize returns of grouping_sets}
\usage{
summarize_rollup(df_list, ...)
}
\arguments{
\item{df_list}{list of 'grouped_df' class}

\item{...}{functions for 'summarize'}
}
\value{
summarized dataframe
}
\description{
S4 method of 'summarize' for class 'grouped_df_list'
}
