% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{with_cube}
\alias{with_cube}
\title{With Cube for R dataframe}
\usage{
with_cube(grouped_df)
}
\arguments{
\item{grouped_df}{'grouped_df' class}
}
\value{
list of dataframes
}
\description{
Compute total amounts at different group levels, producing multiple subtotals. This mirrors the GROUPING SETS operations in SQL.
}
\examples{
mtcars \%>\% group_by(vs, am) \%>\% grouping_sets("vs","am",c("vs","am"))
mtcars \%>\% group_by(vs, am) \%>\% with_rollup() 
mtcars \%>\% group_by(vs, am) \%>\% with_cube() 
}
